+++ Pandora/Paradigm/Structure/Property.hs
@@ -1,4 +1,3 @@
 module Pandora.Paradigm.Structure.Property (module Exports) where
 
 import Pandora.Paradigm.Structure.Property.Nonempty as Exports
-import Pandora.Paradigm.Structure.Property.Hollow as Exports