+++ Pandora/Paradigm/Structure/Binary.hs
@@ -1,4 +1,4 @@
-module Pandora.Paradigm.Structure.Concrete.Binary (Binary, insert) where
+module Pandora.Paradigm.Structure.Binary (Binary, insert) where
 
 import Pandora.Core.Morphism ((&))
 import Pandora.Paradigm.Basis.Wye (Wye (End, Left, Right, Both))