+++ Pandora/Paradigm/Structure/Concrete.hs
@@ -1,5 +1,0 @@
-module Pandora.Paradigm.Structure.Concrete (module Exports) where
-
-import Pandora.Paradigm.Structure.Concrete.Binary as Exports
-import Pandora.Paradigm.Structure.Concrete.Graph as Exports
-import Pandora.Paradigm.Structure.Concrete.Stack as Exports