+++ Pandora/Paradigm/Structure/Property.hs
@@ -1,3 +1,0 @@
-module Pandora.Paradigm.Structure.Property (module Exports) where
-
-import Pandora.Paradigm.Structure.Property.Nonempty as Exports