+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 * Define `Object` instances for `Product` datatype
 
 # 0.1.3
-* Define `Semigroup` and `Monoid` instances for `Cofree` datatype
+* Define `Object` instances for `Cofree` datatype
 * Define all `Object` pattern instances for `Constant` datatype
 * Define `reset` and `shift` methods for `Continuation` datatype
 * Define `Endo` datatype in `Basis` module