+++ CHANGELOG.md
@@ -24,3 +24,4 @@
 * Define `Endo` datatype in `Basis` module
 * Define `Object` instances for transformer schemes
 * Define `Binary` tree concrete structure
+* Define some `Object` instances for `Jack` datatype