+++ CHANGELOG.md
@@ -26,3 +26,4 @@
 * Define `Binary` tree concrete structure
 * Define some `Object` instances for `Jack` datatype
 * Remove `Hollow` ad-hoc typeclass
+* Merge `Property` and `Concrete` modules back